Create BioinformaticsChallenges.Ants.java
Bioinformatic coders warmup challenge
#1
A large number of ants live on an island that reproduce at a rate of 40% per month. There is also an anteater on the island that eats 7000 ants at the end of each month (or all the ants there are if the ant population at that time is less than 7000). When the island's ant population exceeds the maximum of 28000, feeding problems begin, which reduces the growth rate to 31% per month. Write the function count_ants that receives as parameters the number of ants that are on the island at any given time and a number of months. The function must calculate how many ants will be on the island after that number of months. Assume that the population is sampled only at the end of the month and that the anteater only eats at the end of the month.
#A solution in Java: can be run in Command Line/terminal with javac command or use an IDE. 
Package is equivalent to a folder in computer. Make a folder with the above package name and make java class named Ants with above scripts.